Create HomeScreenViewModel for managing todo list data
A `HomeScreenViewModel` is created, utilizing `viewModelScope` to manage the lifecycle of coroutines for retrieving and updating a list of `TodoEntity` objects. It uses `StateFlow` to expose the todo list data and `collect` to update `_list` flow. Hilt is employed for dependency injection, injecting the `TodoDaoRepositoryImpl` for database interactions. The `getAllTodos` method retrieves all todo items.